pyproject.toml
[project]
name = "cert-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "bcrypt==4.0.1",
    "email-validator>=2.2.0",
    "fastapi[standard]>=0.115.12",
    "ipykernel>=6.29.5",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "resend>=2.10.0",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]
